cars/requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
Django==5.0.1
django-rest==0.8.7
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
docker==7.0.0
idna==3.6
packaging==23.2
postgres==4.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
psycopg2-pool==1.1
PyJWT==2.8.0
pytz==2023.4
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
urllib3==2.1.0
